Provisioning and uprovisoing of rack
- added buttons for provision and unprovision

- ssh tunnel to 04 server is tuskar is remote
- eg:
ssh -NfL 16385:127.0.0.1:6385 demo@example.com

if you use non-standards port like me (16385) you can change this
in openstack_dashboard/local/local_settings.py

+++ openstack_dashboard/dashboards/infrastructure/resource_management/racks/forms.py
@@ -125,20 +125,29 @@ class UpdateRackStatus(forms.SelfHandlingForm):
             rack = self.initial.get('rack', None)
             action = request.GET.get('action')
 
-            if action == "start":
-                rack.state = "active"
-            elif action == "reboot":
-                rack.state = "active"
-            elif action == "shutdown":
-                rack.state = "off"
+            if action == "provision":
+                rack = api.tuskar.Rack.provision(
+                    request,
+                    rack.id)
 
-            rack = api.tuskar.Rack.update(
-                request,
-                rack.id,
-                {'state': rack.state}
-            )
+                msg = _('Rack "%s" is being provisioned.') % rack.name
+            else:
+                if action == "start":
+                    rack.state = "active"
+                elif action == "unprovision":
+                    rack.state = "unprovisioned"
+                elif action == "reboot":
+                    rack.state = "active"
+                elif action == "shutdown":
+                    rack.state = "off"
 
-            msg = _('Updated rack "%s" status.') % rack.name
+                rack = api.tuskar.Rack.update(
+                    request,
+                    rack.id,
+                    {'state': rack.state}
+                )
+
+                msg = _('Updated rack "%s" status.') % rack.name
             messages.success(request, msg)
             return True
         except:
